+++ Menu-o-Matic/src/main/groovy/PackMenu.groovy
@@ -25,17 +25,23 @@ class PackMenu{
     static final String   scriptStr    = '_script'
     // static final Boolean  scriptInNote = true
     
-    static final String   titleD1      = 'menu-o-matic'
-    static final String   msgD1        = 'Dialog must show:'
-    static final String[] optionsD1    = ['icons and labels','icons only', 'labels only','cancel']
+    static final String MoMDialogTitle = 'menu-o-matic'
+    static final String iconLabelMsg = 'Dialog must show:'
+    static final String[] optionsD1    = ['icons only', 'labels only','icons and labels','cancel']
     static final String   titleD2      = 'menu-o-matic'
-    static final String   msgD2        = 'Export script code to node\'s :'
+    static final String scriptInMsg = 'Export script code to node\'s :'
     static final String[] optionsD2    = ['note','script1 attribute']
-    static final String   titleD3      = 'menu-o-matic'
-    static final String   msgD3        = 'After clicking a button the focus should:'
-    static final String[] optionsD3    = ['return to mindmap','stay in menu']
+    static final String focusToMapMsg = 'After click return focus to mindmap'
+    //static final String[] optionsD3    = ['return to mindmap','stay in menu']  //focus to map is currently don't needed, I will let the code here just in case in a future version it makes sense again.
     static final String   MoM_TAB_NAME = 'MoM'
-    
+    static final String   textLengthMsg = 'Maximal label text length:'
+    static final String   tabNameMsg    = 'Include toolbar to tab:'
+    static final String   autoLaunchMsg = 'Auto launch toolbar'
+    static final String   permissionsMsg = 'Scripts permissions'
+    static final String   readPermission  = 'WITHOUT_READ_RESTRICTION'
+    static final String   writePermission = 'WITHOUT_WRITE_RESTRICTION'
+    static final String   netPermission   = 'WITHOUT_EXEC_RESTRICTION'
+    static final String   exePermission   = 'WITHOUT_NETWORK_RESTRICTION'
     static final int      maxTextLen   = 50
     static       int      iScript      = 0
 
@@ -50,8 +56,11 @@ class PackMenu{
             focusMap      : 'tbFocusToMap',
             tabName       : 'tbTabName',
             tabIcon       : 'tbTabIcon',
-            color         : 'tbBackColor'
+            color         : 'tbBackColor',
+            permissions   : 'tbPermissions',
+            autoLaunch    : 'tbAutoLaunch'
     ]
+
 
     // endregion: properties
 
@@ -61,25 +70,27 @@ class PackMenu{
         ArrayList icons
         ArrayList labels
         int       maxTextLength
-        Boolean   showIcons
-        Boolean   showLabels
+        boolean   showIcons
+        boolean   showLabels
         String    title
         ArrayList scripts
-        Boolean   focusMap
+        boolean   focusMap //focus to map is currently don't needed, I will let the code here just in case in a future version it makes sense again.
         String    tabName = MoM_TAB_NAME
         String    tabIcon
         String    color
         String    fgColor
+        String    permissions
+        boolean   autoLaunch
 
         public MenuData(Proxy.Node nodoMenu){
 //            if(nodoMenu[TB.title]){
                 this.actions        = nodoMenu[TB.actions]?nodoMenu[TB.actions][1..-2].split(', ').collect{it!='null'?it:null}:[]
                 this.icons          = nodoMenu[TB.icons  ]?nodoMenu[TB.icons  ][1..-2].split(', ').collect{it!='null'?it:null}:[]
                 this.labels         = nodoMenu[TB.labels ]?nodoMenu[TB.labels ][1..-2].split(', ').collect{it!='null'?it:null}:[]
-                this.maxTextLength  = maxTextLen?:50
+                this.maxTextLength  = nodoMenu[TB.maxTextLength].num?:maxTextLen
                 this.showIcons      = nodoMenu[TB.showIcons ].bool
                 this.showLabels     = nodoMenu[TB.showLabels].bool
-                this.focusMap       = nodoMenu[TB.focusMap  ].bool
+                //this.focusMap       = nodoMenu[TB.focusMap  ].bool //focus to map is currently don't needed, I will let the code here just in case in a future version it makes sence again.
                 this.title          = nodoMenu[TB.title]?nodoMenu[TB.title].toString():null
                 def scriptList = []
                 nodoMenu.attributes.names.findAll{it.startsWith(scriptStr)}.each{
@@ -90,10 +101,13 @@ class PackMenu{
                 this.tabIcon        = nodoMenu.icons.first
                 this.color          = nodoMenu.style.backgroundColorSet?nodoMenu.style.backgroundColorCode:null
                 this.fgColor        = nodoMenu.style.textColorSet?nodoMenu.style.textColorCode:null
+                this.permissions    = nodoMenu[TB.permissions]?nodoMenu[TB.permissions].toString():'0001'
+                this.autoLaunch     = nodoMenu[TB.autoLaunch].bool
 //            }
         }
 
-        public MenuData(String title, nAcciones, Boolean showIcons, Boolean showLabels, Boolean focusMap){
+        public MenuData(String title, nAcciones, boolean showIcons, boolean showLabels,
+                        boolean focusMap, int maxTextLength, String tabName, boolean autoLaunch, String permissions){
             iScript      = 0
             this.actions        = nAcciones.collect{accion(it)}
             this.icons          = nAcciones.collect{icono(it)}
@@ -101,7 +115,7 @@ class PackMenu{
             this.maxTextLength  = maxTextLen
             this.showIcons      = showIcons
             this.showLabels     = showLabels
-            this.focusMap       = focusMap
+            //this.focusMap       = focusMap    //focus to map is currently don't needed, I will let the code here just in case in a future version it makes sence again.
             this.title          = title
             def scriptList = []
             def i = 0
@@ -109,6 +123,10 @@ class PackMenu{
                 scriptList << [scriptStr + ++i, WSE.scriptFromNode(it)]
             }
             this.scripts        = scriptList
+            this.maxTextLength  = maxTextLength
+            this.tabName        = tabName
+            this.autoLaunch     = autoLaunch
+            this.permissions     = permissions
         }
 
         public toNode(Proxy.Node nodo){
@@ -118,8 +136,11 @@ class PackMenu{
             nodo[TB.maxTextLength] = this.maxTextLength
             nodo[TB.showIcons]     = this.showIcons
             nodo[TB.showLabels]    = this.showLabels
-            nodo[TB.focusMap]      = this.focusMap
+            //nodo[TB.focusMap]      = this.focusMap   //focus to map is currently don't needed, I will let the code here just in case in a future version it makes sence again.
             nodo[TB.title]         = this.title
+            nodo[TB.tabName]       = this.tabName
+            nodo[TB.permissions]   = this.permissions
+            nodo[TB.autoLaunch]    = this.autoLaunch
             nodo[TB.tabName]       = this.tabName
             if(this.tabIcon){
                 nodo.icons.add(this.tabIcon)
@@ -145,31 +166,43 @@ class PackMenu{
     // region: primary methods
 
     def static getMDfromNodes(nodoBase, boolean useDetails = true){
-        def resp
-        def resp2
-        if(useDetails){
-            def detailsText = nodoBase.details.toString()
-            resp  = optionsD1.find{detailsText.toLowerCase().contains(it)}
-            resp2 = optionsD3.find{detailsText.toLowerCase().contains(it)}
-        }
-        resp    ?= respuestaDialogo(optionsD1,msgD1,titleD1)
-        if(optionsD1[0..2].contains(resp)) {
-            def showIcons  = (resp in optionsD1[0,1])?true:false
-            def showLabels = (resp in optionsD1[0,2])?true:false
-            def nAcciones  = nodoBase.find{it.link?.uri?.scheme == 'menuitem' || WSE.isGroovyNode(it)}
-            resp2          ?= respuestaDialogo(optionsD3,msgD3,titleD3)
-            def focusMap   = ( resp2 == optionsD3[0])
-            return [new MenuData(nodoBase.text, nAcciones, showIcons, showLabels, focusMap), "\n  - $resp\n  - $resp2" ]
+        //get info from child nodes
+        def nAcciones  = nodoBase.find{it.link?.uri?.scheme == 'menuitem' || WSE.isGroovyNode(it)}
+        def hasScripts = nAcciones.any{WSE.isGroovyNode(it)}
+
+        //get info from node
+        def title = nodoBase.text
+
+        //get info from node/dialog
+        def maxTextLength = nodoBase[TB.maxTextLength].num
+        def tabName       = nodoBase[TB.tabName].string
+        def autoLaunch    = nodoBase[TB.autoLaunch]//.bool
+        def showIcons     = nodoBase[TB.showIcons]//.bool
+        def showLabels    = nodoBase[TB.showLabels]//.bool
+        def focusToMap    = nodoBase[TB.focusMap]//.bool
+        def permissions   = nodoBase[TB.permissions].string
+        def forceDialog = false
+        (maxTextLength, tabName, autoLaunch, showIcons, showLabels, focusToMap, permissions) = getConfirmedInfo(forceDialog, hasScripts, maxTextLength, tabName, autoLaunch, showIcons, showLabels, focusToMap, permissions)
+        if(maxTextLength && tabName && autoLaunch!=null && showIcons!=null && showLabels!=null && focusToMap!=null) {
+            return [new MenuData(nodoBase.text, nAcciones, showIcons, showLabels, focusToMap, maxTextLength, tabName, autoLaunch, permissions), "\n  - resp\n  - resp2" ]
         } else {
-            return [null, resp]
+            return [null, 'resp']
         }
     }
-    
+
     def static packMenu(nodoBase){
         def (md, resp) = getMDfromNodes(nodoBase)
         if(md){
             def nodo = nodoBase.createChild(nodoBase.text)
             md.toNode(nodo)
+            if(nodoBase.icons.size()>0)
+                nodo.icons.add(nodoBase.icons.first)
+            def color          = nodoBase.style.backgroundColorSet?nodoBase.style.backgroundColorCode:null
+            if(color)
+                nodo.style.backgroundColorCode = color
+            def fgColor        = nodoBase.style.textColorSet?nodoBase.style.textColorCode:null
+            if(fgColor)
+                nodo.style.textColorCode = fgColor
             nodo.details   = "customMenu  ${resp}"
             nodo.link.text = 'menuitem:_addons.menuOMatic.launchCustomDialog_on_single_node'
             c.statusInfo = 'customMenu node created'
@@ -183,13 +216,29 @@ class PackMenu{
         def scriptInNote
         if (md.scripts){
             def options  = optionsD2
-            def msg      = msgD2
+            def msg      = scriptInMsg
             def title    = titleD2
             def resp     = respuestaDialogo(options,msg,title)
             scriptInNote = (resp == options[0])?true:false
         }
         def nBase = nodoMenu.createChild(md.title)
         nBase.details = nodoMenu.details
+        if(nodoMenu.icons.size()>0)
+            nBase.icons.add(nodoMenu.icons.first)
+        def color          = nodoMenu.style.backgroundColorSet?nodoMenu.style.backgroundColorCode:null
+        if(color)
+            nBase.style.backgroundColorCode = color
+        def fgColor        = nodoMenu.style.textColorSet?nodoMenu.style.textColorCode:null
+        if(fgColor)
+            nBase.style.textColorCode = fgColor
+        nBase[TB.showIcons] = md.showIcons
+        nBase[TB.showLabels] = md.showLabels
+        //nBase[TB.focusMap] = md.focusMap //focus to map is currently don't needed, I will let the code here just in case in a future version it makes sence again.
+        nBase[TB.maxTextLength] = md.maxTextLength
+        nBase[TB.tabName] = md.tabName
+        nBase[TB.autoLaunch] = md.autoLaunch
+        if(md.actions.any{acc -> acc.startsWith(scriptStr)})
+            nBase[TB.permissions] = md.permissions
         md.actions.eachWithIndex{acc, i ->
             def nodo = nBase.createChild(md.labels[i])
             if(acc.startsWith(scriptStr)){
@@ -219,13 +268,65 @@ class PackMenu{
         if (longText) {
             //iconLabel = ic?sl?'icons and labels':'icons only':sl?'labels only':'error'
             iconLabel = ic?sl?optionsD1[0]:optionsD1[1]:sl?optionsD1[2]:'error'
-            //focus = fm?'return to mindmap':'stay in menu'
-            focus = fm?optionsD3[0]:optionsD3[1]
+            focus = '' // fm?optionsD3[0]:optionsD3[1] //focus to map is currently don't needed, I will let the code here just in case in a future version it makes sense again.
         } else {
             iconLabel = ic?sl?'ic+lb':'ic':sl?'lb':'error'
-            focus = fm?'-> map':'-> menu'
+            focus = '' // fm?'-> map':'-> menu' //focus to map is currently don't needed, I will let the code here just in case in a future version it makes sense again.
         }
         return "$title  ($iconLabel, $focus)".toString()
+    }
+
+    def static getConfirmedInfo(forceDialog, hasScripts, maxTextLength, tabName, autoLaunch, showIcons, showLabels, focusToMap, permissions) {
+        //if all info is ready discard inputDialog
+        if(!forceDialog && maxTextLength && tabName && autoLaunch!=null && showIcons!=null && showLabels!=null && focusToMap!=null && (!hasScripts || permissions)){
+            return [maxTextLength, tabName, autoLaunch.bool, showIcons.bool, showLabels.bool, focusToMap.bool, permissions]
+        }
+        //create input objects
+        JTextField textLengthField = new JTextField(maxTextLength?.toString()?:maxTextLen.toString())
+        JTextField tabNameField = new JTextField(tabName?:MoM_TAB_NAME)
+        JCheckBox autoLaunchCheckBox = new JCheckBox(autoLaunchMsg, autoLaunch?autoLaunch.bool:false)
+        JCheckBox focusMapCheckBox = new JCheckBox(focusToMapMsg, focusToMap?focusToMap.bool:true)
+        def selectOption = (((showIcons.bool?1:0) + (showLabels.bool?2:0))?:3) - 1
+        JComboBox iconsLabelsComboBox = new JComboBox(optionsD1.take(3))
+        iconsLabelsComboBox.setSelectedIndex(selectOption)
+        def binPermissions = permissions?Integer.parseInt(permissions, 2):0b0001
+
+        JCheckBox readPermissionCheckBox  = hasScripts? new JCheckBox(readPermission,  (binPermissions & 0b0001)>0): null
+        JCheckBox writePermissionCheckBox = hasScripts? new JCheckBox(writePermission, (binPermissions & 0b0010)>0): null
+        JCheckBox netPermissionCheckBox   = hasScripts? new JCheckBox(netPermission,   (binPermissions & 0b0100)>0): null
+        JCheckBox exePermissionCheckBox   = hasScripts? new JCheckBox(exePermission,   (binPermissions & 0b1000)>0): null
+        //create group component
+        final JComponent[] inputs = new JComponent[] {
+                new JLabel(iconLabelMsg),
+                iconsLabelsComboBox,
+                new JLabel(textLengthMsg),
+                textLengthField,
+                new JLabel(tabNameMsg),
+                tabNameField,
+                autoLaunchCheckBox,
+                //focusMapCheckBox,  //focus to map is currently don't needed, I will let the code here just in case in a future version it makes sence again.
+                hasScripts? new JLabel(permissionsMsg) : null,
+                readPermissionCheckBox,
+                writePermissionCheckBox,
+                netPermissionCheckBox,
+                exePermissionCheckBox,
+        };
+        //get result
+        int result = JOptionPane.showConfirmDialog(ui.getCurrentFrame(), inputs, MoMDialogTitle, JOptionPane.PLAIN_MESSAGE);
+        if (result == JOptionPane.OK_OPTION) {
+            def permiString = hasScripts? Integer.toBinaryString((readPermissionCheckBox.selected?0b0001:0)+(writePermissionCheckBox.selected?0b0010:0)+(netPermissionCheckBox.selected?0b0100:0)+(exePermissionCheckBox.selected?0b1000:0)) : null
+            return [
+                    textLengthField.text?.isInteger()?textLengthField.text.toInteger():maxTextLen,
+                    tabNameField.text?:MoM_TAB_NAME,
+                    autoLaunchCheckBox.isSelected(),
+                    ((iconsLabelsComboBox.getSelectedIndex() + 1) & 0b01)>0,
+                    ((iconsLabelsComboBox.getSelectedIndex() + 1) & 0b10)>0,
+                    focusMapCheckBox.isSelected(),
+                    hasScripts? ('0000' + permiString).takeRight(4) : null,
+            ]
+        } else {
+            return [null, null, null, null, null, null, null]
+        }
     }
 
     // endregion: primary methods